+++ pysatNASA/instruments/cnofs_plp.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
-"""Supports the Planar Langmuir Probe (PLP) onboard the Communication
+"""Module for the C/NOFS PLP instrument.
+
+Supports the Planar Langmuir Probe (PLP) onboard the Communication
 and Navigation Outage Forecasting System (C/NOFS) satellite. Downloads
 data from the NASA Coordinated Data Analysis Web (CDAWeb).
 
@@ -58,10 +60,11 @@ import datetime as dt
 import functools
 import numpy as np
 
-from pysat import logger
 from pysat.instruments.methods import general as mm_gen
-from pysatNASA.instruments.methods import cnofs as mm_cnofs
+
 from pysatNASA.instruments.methods import cdaweb as cdw
+from pysatNASA.instruments.methods import cnofs as mm_cnofs
+from pysatNASA.instruments.methods import general as mm_nasa
 
 # ----------------------------------------------------------------------------
 # Instrument attributes
@@ -79,23 +82,12 @@ _test_dates = {'': {'': dt.datetime(2009, 1, 1)}}
 # ----------------------------------------------------------------------------
 # Instrument methods
 
-
-def init(self):
-    """Initializes the Instrument object with instrument specific values.
-
-    Runs once upon instantiation.
-
-    """
-    logger.info(mm_cnofs.ackn_str)
-    self.acknowledgements = mm_cnofs.ackn_str
-    self.references = '\n'.join((mm_cnofs.refs['mission'],
-                                 mm_cnofs.refs['plp']))
-
-    return
+# Use standard init routine
+init = functools.partial(mm_nasa.init, module=mm_cnofs, name=name)
 
 
 def clean(self):
-    """Routine to return C/NOFS PLP data cleaned to the specified level
+    """Clean C/NOFS PLP data to the specified level.
 
     Note
     ----
@@ -105,7 +97,7 @@ def clean(self):
 
     for key in self.data.columns:
         if key != 'Epoch':
-            fill = self.meta[key, self.meta.labels.fill_val][0]
+            fill = self.meta[key, self.meta.labels.fill_val]
             idx, = np.where(self[key] == fill)
             self[idx, key] = np.nan
     return