+++ pysatNASA/instruments/de2_lang.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
-"""Supports the Langmuir Probe (LANG) instrument on
-Dynamics Explorer 2 (DE2).
+"""Module for the DE2 LANG instrument.
+
+Supports the Langmuir Probe (LANG) instrument on Dynamics Explorer 2 (DE2).
 
 From CDAWeb:
 
@@ -27,8 +28,8 @@ or correct the inflight processed data. Time resolution was 0.5 seconds.
 References
 ----------
 J. P. Krehbiel, L. H. Brace, R. F. Theis, W. H. Pinkus, and R. B. Kaplan,
-The Dynamics Explorer 2 Langmuir Probe (LANG), Space Sci. Instrum., v. 5, n. 4,
-p. 493, 1981.
+"The Dynamics Explorer 2 Langmuir Probe (LANG)", Space Sci. Instrum., 5,
+493-502, 1981.
 
 Properties
 ----------
@@ -47,20 +48,16 @@ Warnings
 - Currently no cleaning routine.
 
 
-Authors
--------
-J. Klenzing
-
 """
 
 import datetime as dt
 import functools
-import warnings
 
-from pysat import logger
 from pysat.instruments.methods import general as mm_gen
-from pysatNASA.instruments.methods import de2 as mm_de2
+
 from pysatNASA.instruments.methods import cdaweb as cdw
+from pysatNASA.instruments.methods import de2 as mm_de2
+from pysatNASA.instruments.methods import general as mm_nasa
 
 # ----------------------------------------------------------------------------
 # Instrument attributes
@@ -79,34 +76,11 @@ _test_dates = {'': {'': dt.datetime(1983, 1, 1)}}
 # Instrument methods
 
 
-def init(self):
-    """Initializes the Instrument object with instrument specific values.
+# Use standard init routine
+init = functools.partial(mm_nasa.init, module=mm_de2, name=name)
 
-    Runs once upon instantiation.
-
-    """
-
-    logger.info(mm_de2.ackn_str)
-    self.acknowledgements = mm_de2.ackn_str
-    self.references = mm_de2.refs['lang']
-    return
-
-
-def clean(self):
-    """Routine to return DE2 LANG data cleaned to the specified level
-
-    Note
-    ----
-    'clean' - Not specified
-    'dusty' - Not specified
-    'dirty' - Not specified
-    'none'  No cleaning applied, routine not called in this case.
-
-    """
-    warnings.warn('No cleaning routines available for DE2 LANG')
-
-    return
-
+# No cleaning, use standard warning function instead
+clean = mm_nasa.clean_warn
 
 # ----------------------------------------------------------------------------
 # Instrument functions