+++ pysatNASA/instruments/methods/de2.py
@@ -1,9 +1,12 @@
 # -*- coding: utf-8 -*-
-"""Provides non-instrument specific routines for DE2 data"""
+"""Provides non-instrument specific routines for DE2 data."""
 
 ackn_str = "The Dynamics Explorer 2 satellite data is provided through CDAWeb"
 
-refs = {'idm': ' '.join(('TBD')),
+refs = {'idm': ' '.join(('Heelis, R. A., W. B. Hanson, C. R. Lippincott, D. R.',
+                         'Zuccaro, L. L. Harmon, B. J. Holt, J. E. Doherty, R.',
+                         'A. Power, The ion drift meter for Dynamics',
+                         'Explorer-B, Space Sci. Instrum., 5, 511, 1981.')),
         'lang': ' '.join(('J. P. Krehbiel, L. H. Brace, R. F. Theis, W. H.',
                           'Pinkus, and R. B. Kaplan, The Dynamics Explorer 2',
                           'Langmuir Probe (LANG), Space Sci. Instrum., v. 5,',
@@ -21,5 +24,8 @@ refs = {'idm': ' '.join(('TBD')),
                           'Hedin, G. R. Carrignan, J. C. Maurer, The',
                           'Dynamics Explorer Wind and Temperature Spectrometer',
                           'Space Sci. Instrum., v. 5, n. 4, p. 417, 1981.')),
-        'vefi': ' '.join(('TBD'))
+        'vefi': ' '.join(('Maynard, N. C., E. A. Bielecki, H. G. Burdick,',
+                          'Instrumentation for vector electric field',
+                          'measurements from DE-B, Space Sci. Instrum., 5,',
+                          '523, 1981.'))
         }